button added

+++ shader.hpp
@@ -1,10 +1,11 @@
 #pragma once
-#include <vector>
 #include <fstream>
 #include <stdexcept>
+#include <vector>
 #include <vulkan/vulkan.h>
 
-static std::vector<char> readFile(const std::string &filename)
+static std::vector<char>
+readFile(const std::string& filename)
 {
     std::ifstream file(filename, std::ios::ate | std::ios::binary);
 
@@ -13,8 +14,8 @@ static std::vector<char> readFile(const std::string &filename)
         throw std::runtime_error("failed to open file!");
     }
 
-    size_t fileSize = (size_t)file.tellg();
-    std::vector<char> buffer(fileSize);
+    i64 fileSize = (i64)file.tellg();
+    std::vector<char> buffer((u64)fileSize);
     file.seekg(0);
     file.read(buffer.data(), fileSize);
     file.close();
@@ -22,16 +23,16 @@ static std::vector<char> readFile(const std::string &filename)
     return buffer;
 }
 
-VkShaderModule createShaderModule(VkDevice device, const std::vector<char> &code)
+VkShaderModule
+createShaderModule(VkDevice device, const std::vector<char>& code)
 {
     VkShaderModuleCreateInfo createInfo{};
     createInfo.sType = VK_STRUCTURE_TYPE_SHADER_MODULE_CREATE_INFO;
     createInfo.codeSize = code.size();
-    createInfo.pCode = reinterpret_cast<const uint32_t *>(code.data());
+    createInfo.pCode = reinterpret_cast<const uint32_t*>(code.data());
 
     VkShaderModule shaderModule;
-    if (vkCreateShaderModule(device, &createInfo, nullptr, &shaderModule) !=
-        VK_SUCCESS)
+    if (vkCreateShaderModule(device, &createInfo, nullptr, &shaderModule) != VK_SUCCESS)
     {
         throw std::runtime_error("failed to create shader module!");
     }
